refactoring for adding new future commands
